Updated templates
After spending a ridiculous amount of time trying to debug the additions
to the MenuItem class, taking a break to work on the templates. Linked
the EAT button to the menu template, started prepping the menu template
for other data. Templates should probably be made more modular so we can
reuse pieces (like the HELP button), and need to pull the CSS styles out
into their own files