Order form
Setting up an Order ModelForm for MenuItems. It doesn't do anything at
the moment, but the goal is to have it either create a new order or add
an item to the existing order. (Maybe we create an order beforehand
though that doesn't contain any menu items? Might make it easier!)